add CRUD operations on Post with right verification

+++ src/main/java/com/hackernews/post/PostController.java
@@ -3,11 +3,15 @@ package com.hackernews.post;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.hackernews.user.CustomUserDetails;
 
 @RestController
 public class PostController {
@@ -25,6 +29,11 @@ public class PostController {
 		return this.postservice.getAllPosts();
 	}
 	
+	@RequestMapping("/posts/{id}")
+	public Post getPost(@PathVariable String id) {
+		return this.postservice.getPost(id);
+	}
+	
 	@RequestMapping(method=RequestMethod.POST , value="/posts")
 	public void addPosts(@RequestBody Post post) {
 		this.postservice.addPost(post);
@@ -36,8 +45,16 @@ public class PostController {
 	}
 	
 	@RequestMapping(method=RequestMethod.DELETE , value="/posts/{id}")
-	public void deletePosts(@PathVariable String id) {
-		this.postservice.deletePost(id);
+	@ResponseBody
+	public String deletePosts(@PathVariable String id, Authentication authentication) {
+		CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+		Post post = this.postservice.getPost(id);
+		if(post.getAuthor().equals(userDetails.getId())) {
+			this.postservice.deletePost(id);
+			return "OK";
+		}
+		return "You don't have the right to delete this post";
+		
 	}
 	
 	@RequestMapping(method=RequestMethod.POST , value="/posts/{id}/comments")
